+++ src/CouplerPanel.cpp
@@ -758,8 +758,17 @@ void CouplerPanel::OnFunctionChange(wxCommandEvent& WXUNUSED(event)) {
 	wxString selectedText = m_functionChoice->GetString(m_functionChoice->GetSelection());
 	m_coupler->setFunction(selectedText);
 	if (m_coupler->getFunction().IsSameAs(wxT("Input"))) {
-		m_availableSwitches->Deselect(m_availableSwitches->GetSelection());
-		m_referencedSwitches->Deselect(m_referencedSwitches->GetSelection());
+		if (m_coupler->getNumberOfSwitches()) {
+			m_coupler->removeAllSwitchReferences();
+			m_referencedSwitches->Clear();
+		}
+		wxArrayInt existingSelections;
+		m_availableSwitches->GetSelections(existingSelections);
+		if (!existingSelections.IsEmpty()) {
+			for (unsigned i = 0; i < existingSelections.GetCount(); i++) {
+				m_availableSwitches->Deselect(existingSelections[i]);
+			}
+		}
 		m_availableSwitches->Enable(false);
 		m_referencedSwitches->Enable(false);
 		m_addReferencedSwitch->Enable(false);
@@ -767,6 +776,12 @@ void CouplerPanel::OnFunctionChange(wxCommandEvent& WXUNUSED(event)) {
 	} else {
 		m_availableSwitches->Enable(true);
 		m_referencedSwitches->Enable(true);
+		if (m_coupler->getFunction().IsSameAs(wxT("NOT"), false) && m_coupler->getNumberOfSwitches() > 1) {
+			while (m_coupler->getNumberOfSwitches() > 1) {
+				m_coupler->removeSwitchReferenceAt(m_coupler->getNumberOfSwitches() - 1);
+				UpdateReferencedSwitches();
+			}
+		}
 	}
 	::wxGetApp().m_frame->m_organ->setModified(true);
 }
@@ -920,7 +935,9 @@ void CouplerPanel::OnAddSwitchReferenceBtn(wxCommandEvent& WXUNUSED(event)) {
 	m_availableSwitches->GetSelections(selectedSwitches);
 	if (!selectedSwitches.IsEmpty()) {
 		for (unsigned i = 0; i < selectedSwitches.GetCount(); i++) {
-			if (!m_coupler->hasSwitchReference(::wxGetApp().m_frame->m_organ->getOrganSwitchAt(selectedSwitches[i]))) {
+			if (!m_coupler->hasSwitchReference(::wxGetApp().m_frame->m_organ->getOrganSwitchAt(selectedSwitches[i])) &&
+				(!(m_coupler->getFunction().IsSameAs(wxT("NOT"), false) && m_coupler->getNumberOfSwitches()))
+			) {
 				m_coupler->addSwitchReference(::wxGetApp().m_frame->m_organ->getOrganSwitchAt(selectedSwitches[i]));
 			}
 		}